% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scInt_New.R
\name{modify_simi}
\alias{modify_simi}
\title{Compute the cell-cell similarities}
\usage{
modify_simi(d_i, d_remained, cluster_labels, dims = 30, k = 5, th = 0.6)
}
\arguments{
\item{d_i}{The dataset i in PCA space.}

\item{d_remained}{Remaining datasets of dataset i in PCA space.}

\item{cluster_labels}{The pre-clustered labels of dataset i.}

\item{dims}{The dimension of PCA. (default 30)}

\item{k}{The number of computed top similarities for each cell. (default 5)}

\item{th}{The similarity threshold for filtering cells in the PCA space. (default 0.6)}
}
\value{
List of identified cell-cell similarities.
}
\description{
Compute the cell-cell similarities
}
