% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scInt.R
\name{normalize.Data}
\alias{normalize.Data}
\title{Normalize scInt object}
\usage{
normalize.Data(object, scale.factor = 10000, do.log = TRUE)
}
\arguments{
\item{object}{\code{scInt} object.}

\item{scale.factor}{The scale factor used for each cell. (default 10000)}

\item{do.log}{Whether do log1p transformation. (default TRUE)}
}
\value{
\code{scInt} object with norm.data slot as normalized data.
}
\description{
This function normalizes the raw dataset of scInt object.
}
