% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scInt.R
\name{run.Integrate}
\alias{run.Integrate}
\title{Perform global integration on multiple normalized datasets}
\usage{
run.Integrate(object, npcs = 40, lambda = 5)
}
\arguments{
\item{object}{\code{scInt} object.}

\item{npcs}{The dimension of projection matrix. (default 40)}

\item{lambda}{The regularization parameter of technical variation. (default 5)}
}
\value{
\code{scInt} object with variation, projection and integrated.data slots set.
}
\description{
This function integrates multiple normalized matrices using identified similarity matrix S.
}
