% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scInt.R
\name{run.pca}
\alias{run.pca}
\title{Compute specific PCA dimensionality reduction for each dataset i}
\usage{
run.pca(norms, i, dims)
}
\arguments{
\item{norms}{The list of normalized matrices.}

\item{i}{The dataset i.}

\item{dims}{The dimension of PCA.}
}
\value{
List of PCA results.
}
\description{
Compute specific PCA dimensionality reduction for each dataset i
}
