% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scInt_New.R
\name{scale_data}
\alias{scale_data}
\title{Scale function}
\usage{
scale_data(
  data.x,
  do.center = T,
  do.scale = T,
  row.means = NULL,
  row.sds = NULL
)
}
\arguments{
\item{data.x}{Input data matrix.}

\item{do.center}{Whether center the row values. (default TRUE)}

\item{do.scale}{Whether scale the row values. (default TRUE)}

\item{row.means}{The provided row means to center. (default NULL)}

\item{row.sds}{The provided row standard deviations to scale. (default NULL)}
}
\description{
This function scales a matrix.
}
