% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scInt.R
\name{compute.Similarity}
\alias{compute.Similarity}
\title{Compute the across-batch cell similarity relationships}
\usage{
compute.Similarity(
  object,
  k = 5,
  dims = 50,
  T_th = 0.6,
  select.T = FALSE,
  cand_Ts = seq(0.5, 0.8, 0.05)
)
}
\arguments{
\item{object}{\code{scInt} object.}

\item{k}{The number of computed top similarities for each cell. (default 5)}

\item{dims}{The dimension of PCA. (default 30)}

\item{T_th}{The similarity threshold for filtering cells in the cPCA space. (default 0.6)}

\item{select.T}{Whether to select parameter T_th. (default FALSE)}

\item{cand_Ts}{The candidate T_th's. (default seq(0.5, 0.8, 0.05))}
}
\value{
\code{scInt} object with S slot set.
}
\description{
Compute the across-batch cell similarity relationships
}
