% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scInt_New.R
\name{run.Integrate}
\alias{run.Integrate}
\title{Perform global integration on multiple normalized datasets}
\usage{
run.Integrate(
  object,
  npcs = 40,
  lambda = 5,
  select.lambda = FALSE,
  cand_lambdas = c(0.1, 1, 5, 10, 15, 20, 50, 100)
)
}
\arguments{
\item{object}{\code{scInt} object.}

\item{npcs}{The dimension of projection matrix. (default 40)}

\item{lambda}{The regularization parameter of technical variation. (default 5)}

\item{select.lambda}{Whether to select parameter lambda. (default FALSE)}

\item{cand_lambdas}{The candidate lambda's. (default c(0.1, 1, 5, 10, 15, 20, 50, 100))}
}
\value{
\code{scInt} object with variation, projection and integrated.data slots set.
}
\description{
This function integrates multiple normalized matrices using identified similarity matrix S.
}
